feat(message): check via twitter msg
